+++ sensor-state/src/sensor_data/sensor_accumulator/temperature_accumulator.rs
@@ -1,14 +1,14 @@
 use crate::sensor_data::SensorData;
 
 pub struct TemperatureAccumulator {
-    pub value: f32,
-    pub max: f32,
-    pub min: Option<f32>
+    pub value: i32,
+    pub max: i32,
+    pub min: Option<i32>
 }
 
 impl TemperatureAccumulator {
     pub const fn default() -> Self {
-        Self { value: 0.0, max: 0.0, min: None }
+        Self { value: 0, max: 0, min: None }
     }
 
     pub fn track(&mut self, _counted_click: bool, sensor_data: &SensorData) {
@@ -27,8 +27,8 @@ impl TemperatureAccumulator {
     }
 
     pub fn clear(&mut self) {
-        self.value = 0.0;
-        self.max = 0.0;
+        self.value = 0;
+        self.max = 0;
         self.min = None
     }
 }
